-1748 - update tests.
