+++ src/test/kotlin/uk/gov/justice/digital/hmpps/hmppsmanageprisonvisitsorchestration/integration/mock/PrisonOffenderSearchMockServer.kt
@@ -5,21 +5,26 @@ import com.github.tomakehurst.wiremock.WireMockServer
 import com.github.tomakehurst.wiremock.client.WireMock.aResponse
 import com.github.tomakehurst.wiremock.client.WireMock.get
 import org.springframework.beans.factory.annotation.Autowired
+import org.springframework.http.HttpStatus
 import org.springframework.http.MediaType
 import uk.gov.justice.digital.hmpps.hmppsmanageprisonvisitsorchestration.dto.prisoner.search.PrisonerDto
 
 class PrisonOffenderSearchMockServer(@Autowired private val objectMapper: ObjectMapper) : WireMockServer(8094) {
-  fun stubGetPrisonerById(prisoner: PrisonerDto) {
+  fun stubGetPrisonerById(offenderNo: String, prisoner: PrisonerDto?) {
+    val responseBuilder = aResponse()
+      .withHeader("Content-Type", MediaType.APPLICATION_JSON_VALUE)
 
     stubFor(
-      get("/prisoner/${prisoner.prisonerNumber}")
+      get("/prisoner/$offenderNo")
         .willReturn(
-          aResponse()
-            .withHeader("Content-Type", MediaType.APPLICATION_JSON_VALUE)
-            .withStatus(200)
-            .withBody(
-              getJsonString(prisoner)
-            )
+          if (prisoner == null) {
+            responseBuilder
+              .withStatus(HttpStatus.NOT_FOUND.value())
+          } else {
+            responseBuilder
+              .withStatus(HttpStatus.OK.value())
+              .withBody(getJsonString(prisoner))
+          }
         )
     )
   }
